refactor: move ts types to separate file
